Zuul status page: Default to 0 instead of ?
I changed this in the recent refactor but it appears the JSON
payload quite often omits the 'queue' properties when they
would be empty (instead of outputting them with an empty
array). Displaying ? makes sense if there is a problem but in
this case it's (unfortunately) normal for Zuul, so lets ignore
the omitted property and fallback to 0.

Change-Id: I65996309588678941d63e5360e9a843ca4a88c88