Zuul status page: Fix flow bug when job doesn't have a url yet.
The status badge is floated to the right. If the job name is too
long to fit on one line, the badge is bumped to the next line.

Normally this is not a problem as the anchor tag clears the
float with `overflow: hidden;`, but when there is no url, there
is no anchor tag.

Fixed by making sure it is wrapped in an element regardless of
the presence of the url.

Also:
- Fixed sample JSON to hit this edge case.
- Separated generic css from the css only applicable in
  bootstrap layout.

Change-Id: I312d9df18d617bd3baca28a5361e5e765163d4d7