Customize Jenkins top left icon
Since we can not reliably change an img src attribute, shrink the image
to 0px and use a padding that will show the background image for the
element.  Rather hacky but works reliably.

The stylesheet jenkins-wmf.css is injected in <head> by a plugin. Url
can be changed in Jenkins settings.

Change-Id: I183b191b3560d85b2da58697fbde8c6f8d2565bc