Revert "Customize Jenkins top left icon"
Per foundation visual identity guidelines, the logo must have space around it.
Also, on a blank background, the regular logo should not be used as it's too
dark. The default Jenkins logo is fine. This is a Jenkins install. It doesn't need
branding I think. Or we can do it good and make a good branding for it, but
as-is the default is better than this.

This reverts commit 0b77026c983a5596b5a9ff2aa0175fdcc63d8aa9.

Change-Id: I2c62d9a40321b6125defe55aa8179fa45f7b6691