zuul: Embed upstream status board instead of forked
This embeds the unmodified status board from upstream
openstack-infra/zuul into our repository.
https://github.com/openstack-infra/zuul/tree/db14a453/etc/status

It is the almost-latest version. Specifically from before this change:

 Status: Don't show jobs for non-live changes
 https://github.com/openstack-infra/zuul/commit/107c385e

Because our installed version of zuul server does not yet
expose the 'item.live' feature in the API and would result in
an incomplete rendering of the dashboard.

And with the following patch applied:

 Status: Expose $.zuul object in zuul_start to allow extension
 https://review.openstack.org/170337

Also bring back behaviour of the following patch that isn't part
of upstream, by using our own init.js.

 * (6314d86f) Refresh graphite graphs together with Zuul status.
   Will be re-based on top.

Change-Id: I04382d874b7b5118b03730bb699d8eb645e66012