Remove use of deprecated jQuery jqXHR `complete` method
An update to the mediawiki.org jQuery library to version 3.0, on which
the Zuul status page relies, caused a deprecation warning to be triggered,
which then broke the javascript that powers the Zuul status page because
the deprecation method logs using mw.log.deprecate(), which does not exist
on this page. We should use a separate copy of jQuery, but for now fix
the breakage by migrating the deprecated method call.

Bug: T174058
Change-Id: I4498e3b63d07d019db51b4f04a1dd20ebc7c1167