zuul: Don't toggle panel when clicking Gerrit patch link
Each patch set panel is clickable in its entirety. Clicks anywhere
on the header of the widget propagate to the wrapper element and
result in the panel being toggled.

This works as expected when clicking on non-interactive areas,
such as the title, ETA, background, or progress bar.

However, this doesn't work as expected when clicking on the
Gerrit patch link, which inadvertently also causes the panel
to be toggled.

Add logic to the click event handler to ignore the event if it
bubbled up from an anchor link.

Test Plan:
* http://localhost/integration-docroot/org/wikimedia/integration/zuul/?demo=basic or:
  /zuul/?source_url=https://integration.wikimedia.org/zuul/status.json
* Click on a Gerrit change link in the top part of a panel.

Previously, it triggers a panel toggle. Now it doesn't.
Clicking elsewhere still toggles the panel.

Change-Id: Ifadf3dfabfc361e1e7d775a62c4c2643b0e2b2c2