cover: Convert coverage sub-nav from breadcrumbs to nav-tabs
Breadcrumbs are mainly intended as a way to go from a subpage back
to an ancestor. Not as a way of going to a subpage.

Given the similarity, it looks somewhat confusing to see the breadcrumbs
visually end with 'MediaWiki extensions' when the user isn't actually
on that page.

Nav-tabs seems like a more appropiate fit within the building blocks
that Bootstrap provides.

Change-Id: Ic7b0cf9d76ed79f12074adf07d054e1fd2456746