zuul: don't retain filter in cookies
It seems rather unexpected that when doing an ad-hoc query, closing
the tab and navigiting to this page again by independent means
a week or two later, that it still starts with it.

Made me think a few times that Zuul was down, when really just
there were no matches for the old filter.

Previously, it was kept in a session cookie. Which in modern
browsers can basically means "forever" (sometimes longer).

instead, keep it in sessionStorage which is associated with the
lifetime of the tab. It does survive explicit "Restore closed tab"
and does survive backward-forward history between it and other
websites. But, does not affect other tabs for the same site, and
does not survive when it is fully closed and re-created days later.

Change-Id: I99d12f856b6861f21c69667178af58ed4656cdca