wmui-osproject: remove dangling caret icon from right side of tile
This was initially left-aligned where it looked better, but made
text not line up vertically which was distracting while scanning
and drew more attention to the button then needed. On the right,
it stands out less but looks off in another way in which is that
it no longer clearly belongs to the button.

I don't think this serves any purpose, so maybe less is more :)
The "…" I think suffices to indicate that it will likely expand
within the page (rather than open a new page) but even if not
that seems optional.

Change-Id: I6b091f7205740981b5dfcc2a9ff62ec0c0a12d58