shared: Update wmui-* to natively use 'wikimedia-ui-base' styles
This makes it easier to update in the future. It adds a very small
perf cost (~ 3KB compressed), which seems tolerable for this
dev-facing portal. Either way it still scores 100/100 on Performance
via a local Lighthouse audit. The whole site is pure HTML/CSS.

The use of native CSS variables means that some colors and lines
fallback to the browser defaults in the ~1-2% of older browsers
that didn't yet support this (e.g. Safari < 9). Thanks to the
semantic HTML we use however, this still produces a very adequate
reading experience, screenshot at:
<https://phabricator.wikimedia.org/F31936555>.

Change-Id: I9a1b27df2f4adb05d045a9aadd06e349ac6f88fc