Support published documents outside of the git checkout
The aim is to have the CI published documentation to be stored outside
of the Apache document root, namely under /srv/doc. That is to separate
concern and is the subject of T149924.  The path will be exposed by
Apache as environment variable WMF_DOC_PATH.

Update the composer scripts start:doc and start:integration to set the
environment variable. That requires tweaking PHP `variables_order` to
have it lookup environment variable (adding `E`).

When Page::getRequestPath() does not find a path in the document root,
have it fallback to lookup under the path pointed by WMF_DOC_ROOT. Add a
test to cover the fallback behavior.

Change the directory index paths for cover, cover-extensions and index
to use "relative" paths. The related classes now look up the given
directory under WMF_DOC_ROOT just like Page::getRequestPath().

Whenever the environment variable is not set or does not resolve to a
realpath(), bails out with an error.

Introduce a web requests router inspired by mediawiki/core which
simulate the Apache directives we have in production namely:

* DirectoryIndex index.html index.php
* Requests to a directory not having those indices use dir.php
* Interpret .php files, anything else is served as it is

Add some boilerplate files and point the PHP built in webserver to it.
That gives us some basic material to check, see README.md.

Remove the obsolete ignore rule in .gitignore since files will be
published outside of this repository checkout.

Bug: T149924
Change-Id: I85e2424c4d18c997d7a903d60a5d47deed6c7ef8