build: Require PHP 7.2, and update PHPCS and PHPUnit to fix PHP 8.0
The new php80 build in CI fails due to composer/semver 1.5.0:
> composer/semver 1.5.0 requires php `^5.3.2 || ^7.0`
> your PHP version (8.0.0) does not satisfy that requirement.

This is a dependency of mediawiki-codesniffer, which has updated
its version of semver in a latest release to fix this issue.

Also update PHPUnit which will otherwise be the next package to
fail since PHPUnit 6 doesn't support PHP 8 per
<https://phpunit.de/supported-versions.html>.

This in turn means bumping the PHP requirement from 7.0 to 7.2,
which is fine since the two production services use 7.2 and 7.3
(per the README), and CI jobs also no longer run on older versions
for this repo.

Change-Id: I4bfdab5719659407a4c99561c3504a18608b2ac3