doc: Support multiple comma-separated languages
Already worked implicitly (although no-one did), but make it official
by making it render nicely if it happens to span multiple lines, as
as well as smoothing over any inconsistent spacing that might be in
the input. The alternative to split/trim/join would be to make the
'lang' key be 'string or array' and then join if needed, but this
seems simpler and matches the way it is displayed which seems useful.

Change-Id: I263ae0c867f62d3c7d40ee8564d6eb58ae50dd04