integration: link to Prometheus dashboard for agents
WMCS instances with Bullseye are no more monitored by Diamond which is
being phased out (it is barely maintained since 2016 and does not
support Python 3).  The new system is based on Prometheus which monitor
any instance.

Change the Grafana board link for Jenkins agents on
https://integration.wikimedia.org/ to the Prometheus based dashboard.

While at it, change the legacy `grafana-labs.wikimedia.org` hostname to
the newish `grafana-cloud.wikimedia.org`.

Bug: T307655
Change-Id: I21f2e616279a2961e37bb56b1dcb28fbdce672b6