zuul: remove symlink indirection
The integration.wikimedia.org/zuul/ page is the last page on this site
that still uses the Bootstrap layout instead of WMUI.

At some point in the past, part of the layout I created was upstreamed
but upstream has since created something different and we also have
a number of added features in our (such as autofocus, and permalinks
for sharing and integrating filters for Gerrit/IRC/Phab).

In the next commit I'll make the code much simpler and in doing so
use WMUI instead of Bootstrap.

Change-Id: I6ede53a9e01584bc1e18820ce9912dd286d2ced8