zuul: Improve separation of concerns in HTML formatting code
== Code quality

Make the HTML formatting code look like an HTML template as much
as possible (which I'll do in the next commit):

* Avoid variables and conditions between HTML chunks, move any
  important logic higher up.
* Use data attributes and dynamic class names to make the logic
  more declarative based on CSS and naturally getting it right,
  instead of inline conditions and such.

Remove reliance on init.js' global state change via $.ajaxSetup().
Insteadm set `cache: false` directly on the one and only $.ajax()
call that this is for.

Reduce use of inline styles and move more of the styles to CSS.

Remove a few unneeded `<span>` elements.

== Design tweaks

* Change "non-voting" to "non<nbsp>voting", to avoid line wrap.
* Reduce width of ETA by using U+2006 half space before hr/min.

== Coding style

* $.each(Array) -> Array.forEach.
* jQuery show()/.hide() -> plain CSS display or HTML hidden attr.
* $.getJSON -> $.ajax().
* String.substr -> String.slice.

Bug: T322168
Change-Id: I06ef6f2c0270f33a3354b9f933799a3b16f77190