zuul: Faster startup, remove delay, remove wrapper, remove app.js
Remove the document-ready wait in zuul.app.js. This isn't needed
because the scripts are loaded from the bottom of the `<body>` which
means the element we need (#zuul_wrapper/#zuul_container) already
exists.

Waiting for document-ready means that we will not render the controls
and will not even start to fetch /status.json until after the rest of
the DOM is ready.

Bug: T322168
Change-Id: I14983d134607acc9119c2d3983a3d6cd84e208fa