doc: Fix handling of /cover/ in router.php
Follows-up 61a94fdae2 which re-ordered the url router such that
doc path is preferred given that PHP8 includes a fallback to index.php
already which otherwise prevents all file browsing.

But... this broke /cover/ which exists in both doc path and web app
at the same time. Between checking for /cover/index.html in the doc path
(which doesn't exist) and serving dir.php file browser, insert a check
to let the web app handle it if it has a /cover/ directory as well.

Change-Id: I249bdbfbf6aaabead55ec9df8cb46b3f7debc588