CoveragePage: Nicer page titles
The buttons have good subject-based labels already:
* MediaWiki extensions
* MediaWiki skins

The page titles were a bit more sentence-y instead like
"MediaWiki extension test coverage". This was mainly for the benefit
of the browser tab (document title) not the H1-heading, since the
in-page heading is in context of the sidebar and layout where we're
in the scope of "Test coverage".

* Adopt those labels as page titles.
* Improve doctitle context to represent "Test coverage".

Bug: T402398
Change-Id: I04d49f190384b946fccef415a83a4cd166fbb4cb